+++ lib/l10n/app_localizations.dart
@@ -62,7 +62,8 @@ import 'app_localizations_pt.dart';
 /// be consistent with the languages listed in the AppLocalizations.supportedLocales
 /// property.
 abstract class AppLocalizations {
-  AppLocalizations(String locale) : localeName = intl.Intl.canonicalizedLocale(locale.toString());
+  AppLocalizations(String locale)
+    : localeName = intl.Intl.canonicalizedLocale(locale.toString());
 
   final String localeName;
 
@@ -70,7 +71,8 @@ abstract class AppLocalizations {
     return Localizations.of<AppLocalizations>(context, AppLocalizations);
   }
 
-  static const LocalizationsDelegate<AppLocalizations> delegate = _AppLocalizationsDelegate();
+  static const LocalizationsDelegate<AppLocalizations> delegate =
+      _AppLocalizationsDelegate();
 
   /// A list of this localizations delegate along with the default localizations
   /// delegates.
@@ -82,18 +84,19 @@ abstract class AppLocalizations {
   /// Additional delegates can be added by appending to this list in
   /// MaterialApp. This list does not have to be used at all if a custom list
   /// of delegates is preferred or required.
-  static const List<LocalizationsDelegate<dynamic>> localizationsDelegates = <LocalizationsDelegate<dynamic>>[
-    delegate,
-    GlobalMaterialLocalizations.delegate,
-    GlobalCupertinoLocalizations.delegate,
-    GlobalWidgetsLocalizations.delegate,
-  ];
+  static const List<LocalizationsDelegate<dynamic>> localizationsDelegates =
+      <LocalizationsDelegate<dynamic>>[
+        delegate,
+        GlobalMaterialLocalizations.delegate,
+        GlobalCupertinoLocalizations.delegate,
+        GlobalWidgetsLocalizations.delegate,
+      ];
 
   /// A list of this localizations delegate's supported locales.
   static const List<Locale> supportedLocales = <Locale>[
     Locale('en'),
     Locale('pt'),
-    Locale('pt', 'BR')
+    Locale('pt', 'BR'),
   ];
 
   /// No description provided for @myPantry.
@@ -1943,6 +1946,54 @@ abstract class AppLocalizations {
   /// In en, this message translates to:
   /// **'Reminds you to add products regularly'**
   String get inactivityReminderChannelDescription;
+
+  /// No description provided for @weeklyRecipeSuggestionTitle.
+  ///
+  /// In en, this message translates to:
+  /// **'Recipe idea from your pantry'**
+  String get weeklyRecipeSuggestionTitle;
+
+  /// No description provided for @weeklyRecipeSuggestionBody.
+  ///
+  /// In en, this message translates to:
+  /// **'How about {recipeName}? It matches {itemCount} ingredients you have. Always check ingredients for dietary suitability.'**
+  String weeklyRecipeSuggestionBody(String recipeName, int itemCount);
+
+  /// No description provided for @weeklyRecipeSuggestionEnabled.
+  ///
+  /// In en, this message translates to:
+  /// **'Weekly recipe suggestions'**
+  String get weeklyRecipeSuggestionEnabled;
+
+  /// No description provided for @weeklyRecipeSuggestionDay.
+  ///
+  /// In en, this message translates to:
+  /// **'Suggestion day'**
+  String get weeklyRecipeSuggestionDay;
+
+  /// No description provided for @weeklyRecipeSuggestionTime.
+  ///
+  /// In en, this message translates to:
+  /// **'Suggestion time'**
+  String get weeklyRecipeSuggestionTime;
+
+  /// No description provided for @weeklyRecipeSuggestionSet.
+  ///
+  /// In en, this message translates to:
+  /// **'Weekly recipe suggestion updated.'**
+  String get weeklyRecipeSuggestionSet;
+
+  /// No description provided for @weeklyRecipeSuggestionChannelName.
+  ///
+  /// In en, this message translates to:
+  /// **'Recipe suggestions'**
+  String get weeklyRecipeSuggestionChannelName;
+
+  /// No description provided for @weeklyRecipeSuggestionChannelDescription.
+  ///
+  /// In en, this message translates to:
+  /// **'Weekly recipe ideas from your pantry'**
+  String get weeklyRecipeSuggestionChannelDescription;
 
   /// No description provided for @notificationDeniedWarning.
   ///
@@ -4093,7 +4144,8 @@ abstract class AppLocalizations {
   String unitSystemChanged(String system);
 }
 
-class _AppLocalizationsDelegate extends LocalizationsDelegate<AppLocalizations> {
+class _AppLocalizationsDelegate
+    extends LocalizationsDelegate<AppLocalizations> {
   const _AppLocalizationsDelegate();
 
   @override
@@ -4102,34 +4154,38 @@ class _AppLocalizationsDelegate extends LocalizationsDelegate<AppLocalizations> 
   }
 
   @override
-  bool isSupported(Locale locale) => <String>['en', 'pt'].contains(locale.languageCode);
+  bool isSupported(Locale locale) =>
+      <String>['en', 'pt'].contains(locale.languageCode);
 
   @override
   bool shouldReload(_AppLocalizationsDelegate old) => false;
 }
 
 AppLocalizations lookupAppLocalizations(Locale locale) {
-
   // Lookup logic when language+country codes are specified.
   switch (locale.languageCode) {
-    case 'pt': {
-  switch (locale.countryCode) {
-    case 'BR': return AppLocalizationsPtBr();
-   }
-  break;
-   }
+    case 'pt':
+      {
+        switch (locale.countryCode) {
+          case 'BR':
+            return AppLocalizationsPtBr();
+        }
+        break;
+      }
   }
 
   // Lookup logic when only language code is specified.
   switch (locale.languageCode) {
-    case 'en': return AppLocalizationsEn();
-    case 'pt': return AppLocalizationsPt();
+    case 'en':
+      return AppLocalizationsEn();
+    case 'pt':
+      return AppLocalizationsPt();
   }
 
   throw FlutterError(
     'AppLocalizations.delegate failed to load unsupported locale "$locale". This is likely '
     'an issue with the localizations generation tool. Please file an issue '
     'on GitHub with a reproducible sample app and the gen-l10n configuration '
-    'that was used.'
+    'that was used.',
   );
 }